requirements.txt
django
djangorestframework
djoser
djangorestframework_simplejwt
python-decouple